implementing sqlite to the database
